+++ src/ts/markdown/codeRender.ts
@@ -7,9 +7,9 @@ export const codeRender = (element: HTMLElement, lang: keyof II18n = "zh_CN") =>
             e.parentElement.classList.contains("vditor-ir__marker--pre ")) {
             return;
         }
-        if (e.classList.contains("language-mermaid") || e.classList.contains("language-echarts") ||
-            e.classList.contains("language-mindmap") || e.classList.contains("language-abc") ||
-            e.classList.contains("language-graphviz")) {
+        if (e.classList.contains("language-mermaid") || e.classList.contains("language-flowchart") ||
+            e.classList.contains("language-echarts") || e.classList.contains("language-mindmap") ||
+            e.classList.contains("language-abc") || e.classList.contains("language-graphviz")) {
             return;
         }